refactor admin and user page to use centralized axios instance
