+++ src/customer/components/ProtectedRoute.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { Navigate, useLocation } from "react-router-dom";
-import AuthContext from "../context/AuthContext";
+import AuthContext from "./context/AuthContext";
 
 const ProtectedRoute = ({ children, role }) => {
   const { user, token } = useContext(AuthContext);
